+++ examples/polymorphic/expected/Operations/AllMembers/AllMembers.php
@@ -37,4 +37,9 @@ class AllMembers extends \Spawnia\Sailor\ObjectLike
             '__typename' => new \Spawnia\Sailor\Convert\NonNullConverter(new \Spawnia\Sailor\Convert\StringConverter),
         ];
     }
+
+    public static function endpoint(): string
+    {
+        return 'polymorphic';
+    }
 }