+++ examples/polymorphic/expected/Operations/UserOrPost/UserOrPost.php
@@ -38,4 +38,9 @@ class UserOrPost extends \Spawnia\Sailor\ObjectLike
             '__typename' => new \Spawnia\Sailor\Convert\NonNullConverter(new \Spawnia\Sailor\Convert\StringConverter),
         ];
     }
+
+    public static function endpoint(): string
+    {
+        return 'polymorphic';
+    }
 }