+++ src/Codegen/OperationGenerator.php
@@ -352,7 +352,8 @@ class OperationGenerator implements ClassGenerator
     {
         return new ObjectLikeBuilder(
             $name,
-            $this->currentNamespace()
+            $this->currentNamespace(),
+            $this->endpointName
         );
     }