+++ src/Result.php
@@ -3,12 +3,13 @@
 namespace Spawnia\Sailor;
 
 use Spawnia\Sailor\Error\Error;
+use Spawnia\Sailor\Error\ResultErrorsException;
 use stdClass;
 
 /**
  * @property \Spawnia\Sailor\ObjectLike|null $data The result of executing the requested operation.
  */
-abstract class Result
+abstract class Result implements BelongsToEndpoint
 {
     /**
      * A non‐empty list of errors or `null` if there are no errors.
@@ -36,7 +37,7 @@ abstract class Result
     abstract public function errorFree(): ErrorFreeResult;
 
     /**
-     * The configured endpoint the result belongs to.
+     * The configured endpoint this class belongs to.
      */
     abstract public static function endpoint(): string;
 
@@ -48,7 +49,7 @@ abstract class Result
         $instance = new static();
 
         $instance->errors = isset($response->errors)
-            ? array_map([Configuration::endpoint(static::endpoint()), 'parseError'], $response->errors)
+            ? static::parseErrors($response->errors)
             : null;
         $instance->extensions = $response->extensions ?? null;
 
@@ -81,22 +82,44 @@ abstract class Result
     public static function fromErrors(array $errors): self
     {
         $instance = new static();
-        $instance->errors = array_map([Configuration::endpoint(static::endpoint()), 'parseError'], $errors);
+        $instance->errors = static::parseErrors($errors);
 
         return $instance;
     }
 
     /**
+     * @param array<int, stdClass> $errors
+     *
+     * @return array<int, Error>
+     */
+    protected static function parseErrors(array $errors): array
+    {
+        $endpoint = Configuration::endpoint(static::endpoint());
+
+        return array_map(
+            static function (stdClass $raw) use ($endpoint): Error {
+                $parsed = $endpoint->parseError($raw);
+                $parsed->endpointName = static::endpoint();
+
+                return $parsed;
+            },
+            $errors
+        );
+    }
+
+    /**
      * Throw an exception if errors are present in the result.
      *
-     * @throws \Spawnia\Sailor\ResultErrorsException
+     * @throws \Spawnia\Sailor\Error\ResultErrorsException
      *
      * @return $this
      */
     public function assertErrorFree(): self
     {
         if (isset($this->errors)) {
-            throw new ResultErrorsException($this->errors);
+            $exception = new ResultErrorsException($this->errors, static::endpoint());
+
+            throw $exception;
         }
 
         return $this;