+++ CHANGELOG.md
@@ -7,6 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## v0.16.0
+
+### Added
+
+- Add configuration `EndpointConfig::errorsAreClientSafe()` to propagate client-safety of endpoint errors
+
+### Changed
+
+- Move `Spawnia\Sailor\ResultErrorsException` to `Spawnia\Sailor\Error\ResultErrorsException`
+- Move `Spawnia\Sailor\InvalidDataException` to `Spawnia\Sailor\Error\InvalidDataException`
+- Include only messages in `ResultErrorsException::$message`, expose full `Error` objects as `ResultErrorsException::$errors`
+
+### Removed
+
+- Remove `Spawnia\Sailor\Response::assertErrorFree()`
+
 ## v0.15.0
 
 ### Added